// src/clients/generated/authClient/models/authResponse.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * order-manager
 * API order-manager
 * OpenAPI spec version: 1.0.0
 */

export interface AuthResponse {
  /**
   * Токен доступа
   * @maxLength 255
   */
  accessToken?: string;
  /**
   * Токен обновления
   * @maxLength 255
   */
  refreshToken?: string;
  /**
   * Идентификатор пользователя
   * @maxLength 36
   */
  userId?: string;
}